pgkernelgateway/src/test/javascript/spec/app/entities/person-type/person-type-update.component.spec.ts
/* tslint:disable max-line-length */
import { ComponentFixture, TestBed, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { FormBuilder } from '@angular/forms';
import { Observable, of } from 'rxjs';

import { PgkernelgatewayTestModule } from '../../../test.module';
import { PersonTypeUpdateComponent } from 'app/entities/person-type/person-type-update.component';
import { PersonTypeService } from 'app/entities/person-type/person-type.service';
import { PersonType } from 'app/shared/model/person-type.model';

describe('Component Tests', () => {
  describe('PersonType Management Update Component', () => {
    let comp: PersonTypeUpdateComponent;
    let fixture: ComponentFixture<PersonTypeUpdateComponent>;
    let service: PersonTypeService;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [PgkernelgatewayTestModule],
        declarations: [PersonTypeUpdateComponent],
        providers: [FormBuilder]
      })
        .overrideTemplate(PersonTypeUpdateComponent, '')
        .compileComponents();

      fixture = TestBed.createComponent(PersonTypeUpdateComponent);
      comp = fixture.componentInstance;
      service = fixture.debugElement.injector.get(PersonTypeService);
    });

    describe('save', () => {
      it('Should call update service on save for existing entity', fakeAsync(() => {
        // GIVEN
        const entity = new PersonType(123);
        spyOn(service, 'update').and.returnValue(of(new HttpResponse({ body: entity })));
        comp.updateForm(entity);
        // WHEN
        comp.save();
        tick(); // simulate async

        // THEN
        expect(service.update).toHaveBeenCalledWith(entity);
        expect(comp.isSaving).toEqual(false);
      }));

      it('Should call create service on save for new entity', fakeAsync(() => {
        // GIVEN
        const entity = new PersonType();
        spyOn(service, 'create').and.returnValue(of(new HttpResponse({ body: entity })));
        comp.updateForm(entity);
        // WHEN
        comp.save();
        tick(); // simulate async

        // THEN
        expect(service.create).toHaveBeenCalledWith(entity);
        expect(comp.isSaving).toEqual(false);
      }));
    });
  });
});
